Refactor update logic into modules
