--- src/update.rs
@@ -1,2922 +1,0 @@
-use std::ffi::OsString;
-use std::path::{Path, PathBuf};
-use std::process::{Command, Stdio};
-use std::time::{Duration, Instant, SystemTime, UNIX_EPOCH};
-
-use flate2::read::GzDecoder;
-use http::header::{ACCEPT, CONTENT_LENGTH, LOCATION, USER_AGENT};
-use http::{HeaderMap, HeaderValue, Method, StatusCode};
-use http_body_util::BodyExt;
-use serde::{Deserialize, Serialize};
-use sha2::{Digest, Sha256};
-use time::OffsetDateTime;
-use time::format_description::well_known::Rfc3339;
-use tokio::io::AsyncWriteExt;
-use tokio::sync::{Mutex, mpsc};
-
-use crate::cli::Cli;
-use crate::core;
-use crate::duration::{TimeoutBudget, duration_from_seconds, format_go_duration};
-use crate::error::{FetchError, write_warning_with_separator_with_color};
-use crate::fileutil::FileLock;
-use crate::http::{client, transport};
-use crate::output::progress::{self, BarCounter, ProgressPrinter, SpinnerCounter};
-
-#[cfg(any(windows, test))]
-const RELOCATED_SUFFIX: &str = ".__relocated.exe";
-#[cfg(any(windows, test))]
-const SELF_DELETE_SUFFIX: &str = ".__selfdelete.exe";
-#[cfg(any(windows, test))]
-const TEMP_EXE_SUFFIX: &str = ".__temp.exe";
-#[cfg(windows)]
-const SELF_DELETE_ENV: &str = "FETCH_INTERNAL_UPDATE_SELF_DELETE";
-const INTERNAL_UPDATE_URL_ENV: &str = "FETCH_INTERNAL_UPDATE_URL";
-const MAX_UPDATE_ARTIFACT_BYTES: u64 = 128 * 1024 * 1024;
-const MAX_UPDATE_CHECKSUM_BYTES: u64 = 1024;
-const MAX_UPDATE_UNPACKED_BYTES: u64 = MAX_UPDATE_ARTIFACT_BYTES * 4;
-const MAX_UPDATE_ARCHIVE_ENTRIES: usize = 128;
-const UPDATE_LOCK_WAIT_TIMEOUT: Duration = Duration::from_secs(30);
-
-#[derive(Debug, Deserialize)]
-struct Release {
-    tag_name: String,
-    assets: Vec<Asset>,
-}
-
-#[derive(Debug, Deserialize)]
-struct Asset {
-    name: String,
-    browser_download_url: String,
-}
-
-#[derive(Debug, PartialEq, Eq)]
-struct ReleaseArtifact<'a> {
-    archive_name: &'a str,
-    archive_url: &'a str,
-    checksum_url: &'a str,
-}
-
-struct UpdateClient<'a> {
-    cli: Option<&'a Cli>,
-    timeout: Option<Duration>,
-    connect_timeout: Option<Duration>,
-    allow_insecure_http: bool,
-    client: Mutex<Option<UpdateCachedClient>>,
-}
-
-#[derive(Clone)]
-struct UpdateCachedClient {
-    origin: String,
-    client: client::UrlClient,
-}
-
-pub async fn execute(cli: &Cli) -> Result<i32, FetchError> {
-    crate::tls::install_default_crypto_provider();
-    let client = UpdateClient::new(cli)?;
-
-    let cache_dir = cache_dir()?;
-    let lock_timeout = update_lock_wait_timeout(client.timeout);
-    let _lock = acquire_update_lock_with_timeout(
-        &cache_dir,
-        true,
-        cli.silent,
-        cli.color.as_deref(),
-        lock_timeout,
-    )?
-    .ok_or_else(|| FetchError::Message("unable to acquire update lock".to_string()))?;
-    let result = update_inner(&client, cli.silent, cli.dry_run).await;
-    record_last_attempt_time(&cache_dir);
-    result?;
-    Ok(0)
-}
-
-pub(crate) fn maybe_spawn_auto_update(value: &str, config_path: Option<&Path>) {
-    let Some(interval) = parse_auto_update_interval(value) else {
-        return;
-    };
-    if should_attempt_update(interval).is_ok_and(|ok| !ok) {
-        return;
-    }
-
-    let Ok(path) = current_exe() else {
-        return;
-    };
-    let mut command = Command::new(path);
-    command
-        .args(auto_update_command_args(config_path))
-        .stdin(Stdio::null())
-        .stdout(Stdio::null())
-        .stderr(Stdio::null());
-    if std::env::var_os("FETCH_INTERNAL_SYNC_AUTO_UPDATE").is_some() {
-        let _ = command.status();
-    } else {
-        let _ = command.spawn();
-    }
-}
-
-fn auto_update_command_args(config_path: Option<&Path>) -> Vec<OsString> {
-    let mut args = Vec::new();
-    if let Some(path) = config_path {
-        push_arg(&mut args, "--config");
-        args.push(path.as_os_str().to_os_string());
-    }
-    push_arg(&mut args, "--update");
-    push_arg(&mut args, "--timeout=300");
-    push_arg(&mut args, "--silent");
-    args
-}
-
-fn push_arg(args: &mut Vec<OsString>, value: impl Into<OsString>) {
-    args.push(value.into());
-}
-
-#[cfg(windows)]
-pub fn maybe_run_self_delete_helper() -> Option<i32> {
-    let data = std::env::var(SELF_DELETE_ENV).ok()?;
-    let exe_path = std::env::current_exe().ok()?;
-    if !path_has_suffix(&exe_path, SELF_DELETE_SUFFIX) {
-        return None;
-    }
-
-    match run_self_delete_helper(&data) {
-        Ok(()) => Some(0),
-        Err(_) => Some(1),
-    }
-}
-
-#[cfg(not(windows))]
-pub fn maybe_run_self_delete_helper() -> Option<i32> {
-    None
-}
-
-async fn update_inner(
-    client: &UpdateClient<'_>,
-    silent: bool,
-    dry_run: bool,
-) -> Result<(), FetchError> {
-    let exe_path = current_exe()?;
-    if !can_replace_file(&exe_path) {
-        return Err(format!(
-            "the current process does not have write permission to '{}'",
-            exe_path.display()
-        )
-        .into());
-    }
-    let version = core::version();
-
-    write_msg(silent, "Fetching latest release...\n");
-    let latest = latest_release(client).await?;
-
-    if latest.tag_name == version {
-        if !silent {
-            eprintln!("Already using the latest version ({}).", latest.tag_name);
-        }
-        return Ok(());
-    }
-
-    if dry_run {
-        if !silent {
-            eprintln!("Update available: {version} -> {}", latest.tag_name);
-        }
-        return Ok(());
-    }
-
-    let release_artifact = release_artifact(&latest).ok_or_else(|| {
-        FetchError::Message(format!(
-            "no release artifact and checksum found for {}/{}",
-            goos(),
-            goarch()
-        ))
-    })?;
-    if !silent {
-        eprintln!("Downloading {}\n", latest.tag_name);
-    }
-
-    let checksum = download_checksum(client, release_artifact.checksum_url).await?;
-
-    let temp_dir = create_update_temp_dir()?;
-    let unpack_dir = temp_dir.path().join("unpacked");
-    std::fs::create_dir(&unpack_dir)?;
-    let actual_checksum = download_and_unpack_artifact(
-        client,
-        release_artifact.archive_name,
-        release_artifact.archive_url,
-        &unpack_dir,
-        temp_dir.path(),
-        silent,
-    )
-    .await?;
-    verify_artifact_checksum(release_artifact.archive_name, &actual_checksum, &checksum)?;
-
-    let src = unpack_dir.join(fetch_filename());
-    let replace_result = self_replace(&exe_path, &src);
-    replace_result?;
-
-    write_update_success(silent, version, &latest.tag_name);
-    Ok(())
-}
-
-async fn latest_release(client: &UpdateClient<'_>) -> Result<Release, FetchError> {
-    let url = format!(
-        "{}/repos/ryanfowler/fetch/releases/latest",
-        update_url().trim_end_matches('/')
-    );
-    let response = update_get(client, &url).await?;
-    if !response.status().is_success() {
-        return Err(format!(
-            "unable to fetch the latest release: received status: {}",
-            response.status().as_u16()
-        )
-        .into());
-    }
-    let release: Release = serde_json::from_slice(response.body())
-        .map_err(|err| FetchError::Message(format!("unable to fetch the latest release: {err}")))?;
-    if release.tag_name.is_empty() {
-        return Err("unable to fetch the latest release: no tag found".into());
-    }
-    Ok(release)
-}
-
-async fn download_and_unpack_artifact(
-    client: &UpdateClient<'_>,
-    archive_name: &str,
-    artifact_url: &str,
-    unpack_dir: &Path,
-    scratch_dir: &Path,
-    silent: bool,
-) -> Result<String, FetchError> {
-    download_and_unpack_artifact_with_limit(
-        client,
-        archive_name,
-        artifact_url,
-        unpack_dir,
-        scratch_dir,
-        silent,
-        MAX_UPDATE_ARTIFACT_BYTES,
-    )
-    .await
-}
-
-async fn download_and_unpack_artifact_with_limit(
-    client: &UpdateClient<'_>,
-    archive_name: &str,
-    artifact_url: &str,
-    unpack_dir: &Path,
-    scratch_dir: &Path,
-    silent: bool,
-    max_artifact_bytes: u64,
-) -> Result<String, FetchError> {
-    if archive_name.ends_with(".zip") {
-        let archive_path = scratch_dir.join("artifact.zip");
-        let actual = download_artifact_to_file_with_limit(
-            client,
-            artifact_url,
-            &archive_path,
-            silent,
-            max_artifact_bytes,
-        )
-        .await?;
-        unpack_zip_artifact_from_file(unpack_dir, &archive_path)?;
-        Ok(actual)
-    } else if archive_name.ends_with(".tar.gz") || archive_name.ends_with(".tgz") {
-        download_and_unpack_tar_gz_artifact_with_limit(
-            client,
-            artifact_url,
-            unpack_dir,
-            silent,
-            max_artifact_bytes,
-        )
-        .await
-    } else {
-        Err(format!("unsupported self-update archive format: {archive_name}").into())
-    }
-}
-
-async fn download_artifact_to_file_with_limit(
-    client: &UpdateClient<'_>,
-    artifact_url: &str,
-    archive_path: &Path,
-    silent: bool,
-    max_artifact_bytes: u64,
-) -> Result<String, FetchError> {
-    let response = update_get_stream(client, artifact_url).await?;
-    let content_length = validate_artifact_response(&response, max_artifact_bytes)?;
-    let mut progress = UpdateDownloadProgress::maybe_start(silent, content_length);
-    let result =
-        write_artifact_response_to_file(response, archive_path, &mut progress, max_artifact_bytes)
-            .await;
-    progress.finish();
-    result
-}
-
-async fn download_and_unpack_tar_gz_artifact_with_limit(
-    client: &UpdateClient<'_>,
-    artifact_url: &str,
-    unpack_dir: &Path,
-    silent: bool,
-    max_artifact_bytes: u64,
-) -> Result<String, FetchError> {
-    let response = update_get_stream(client, artifact_url).await?;
-    let content_length = validate_artifact_response(&response, max_artifact_bytes)?;
-    let mut progress = UpdateDownloadProgress::maybe_start(silent, content_length);
-    let result =
-        unpack_tar_gz_response(response, unpack_dir, &mut progress, max_artifact_bytes).await;
-    progress.finish();
-    result
-}
-
-fn validate_artifact_response(
-    response: &UpdateStreamingResponse,
-    max_artifact_bytes: u64,
-) -> Result<i64, FetchError> {
-    if !response.status().is_success() {
-        return Err(format!(
-            "fetching artifact: downloading artifact: received status: {}",
-            response.status().as_u16()
-        )
-        .into());
-    }
-
-    if let Some(len) = response.content_length()
-        && len > max_artifact_bytes
-    {
-        return Err(format!("update artifact is too large: {len} bytes").into());
-    }
-
-    Ok(response
-        .content_length()
-        .and_then(|value| i64::try_from(value).ok())
-        .unwrap_or(-1))
-}
-
-async fn write_artifact_response_to_file(
-    response: UpdateStreamingResponse,
-    archive_path: &Path,
-    progress: &mut UpdateDownloadProgress,
-    max_artifact_bytes: u64,
-) -> Result<String, FetchError> {
-    let mut file = tokio::fs::File::create(archive_path).await?;
-    let mut hasher = Sha256::new();
-    let mut received = 0u64;
-    let budget = response.budget;
-    let (mut body, _) = response.response.into_body_with_deadline();
-
-    while let Some(data) = next_update_body_data(budget, &mut body).await? {
-        account_artifact_chunk(
-            &mut received,
-            data.len(),
-            max_artifact_bytes,
-            "update artifact exceeded maximum allowed size",
-        )?;
-        hasher.update(&data);
-        progress.add(data.len());
-        file.write_all(&data).await?;
-    }
-    file.flush().await?;
-
-    Ok(finish_sha256_hex(hasher))
-}
-
-async fn unpack_tar_gz_response(
-    response: UpdateStreamingResponse,
-    unpack_dir: &Path,
-    progress: &mut UpdateDownloadProgress,
-    max_artifact_bytes: u64,
-) -> Result<String, FetchError> {
-    let (tx, rx) = mpsc::channel::<bytes::Bytes>(8);
-    let unpack_dir = unpack_dir.to_path_buf();
-    let unpack_task = tokio::task::spawn_blocking(move || {
-        unpack_tar_gz_artifact_from_reader(
-            &unpack_dir,
-            ChannelReader::new(rx),
-            ArchiveExtractionLimits::default(),
-        )
-    });
-
-    let mut hasher = Sha256::new();
-    let mut received = 0u64;
-    let budget = response.budget;
-    let (mut body, _) = response.response.into_body_with_deadline();
-
-    while let Some(data) = match next_update_body_data(budget, &mut body).await {
-        Ok(data) => data,
-        Err(err) => {
-            drop(tx);
-            let _ = finish_unpack_task(unpack_task).await;
-            return Err(err);
-        }
-    } {
-        if let Err(err) = account_artifact_chunk(
-            &mut received,
-            data.len(),
-            max_artifact_bytes,
-            "update artifact exceeded maximum allowed size",
-        ) {
-            drop(tx);
-            let _ = finish_unpack_task(unpack_task).await;
-            return Err(err);
-        }
-        hasher.update(&data);
-        progress.add(data.len());
-        if tx.send(data).await.is_err() {
-            finish_unpack_task(unpack_task).await?;
-            return Err(FetchError::Message(
-                "self-update archive unpack task ended before download completed".to_string(),
-            ));
-        }
-    }
-
-    drop(tx);
-    finish_unpack_task(unpack_task).await?;
-    Ok(finish_sha256_hex(hasher))
-}
-
-async fn next_update_body_data(
-    budget: TimeoutBudget,
-    body: &mut transport::Body,
-) -> Result<Option<bytes::Bytes>, FetchError> {
-    while let Some(frame) = budget
-        .run(Box::pin(async {
-            match body.frame().await {
-                Some(Ok(frame)) => Ok(Some(frame)),
-                Some(Err(err)) => Err(FetchError::Runtime(format!("response body error: {err}"))),
-                None => Ok(None),
-            }
-        }))
-        .await?
-    {
-        if let Ok(data) = frame.into_data()
-            && !data.is_empty()
-        {
-            return Ok(Some(data));
-        }
-    }
-    Ok(None)
-}
-
-fn account_artifact_chunk(
-    received: &mut u64,
-    chunk_len: usize,
-    max_artifact_bytes: u64,
-    limit_error: &'static str,
-) -> Result<(), FetchError> {
-    let chunk = u64::try_from(chunk_len).unwrap_or(u64::MAX);
-    if received.saturating_add(chunk) > max_artifact_bytes {
-        return Err(limit_error.into());
-    }
-    *received = received.saturating_add(chunk);
-    Ok(())
-}
-
-async fn finish_unpack_task(
-    task: tokio::task::JoinHandle<Result<(), FetchError>>,
-) -> Result<(), FetchError> {
-    task.await.map_err(|err| {
-        FetchError::Runtime(format!("self-update archive unpack task failed: {err}"))
-    })?
-}
-
-async fn download_checksum(
-    client: &UpdateClient<'_>,
-    checksum_url: &str,
-) -> Result<String, FetchError> {
-    let response = update_get_stream(client, checksum_url).await?;
-    if !response.status().is_success() {
-        return Err(format!(
-            "fetching artifact checksum: received status: {}",
-            response.status().as_u16()
-        )
-        .into());
-    }
-
-    if let Some(len) = response.content_length()
-        && len > MAX_UPDATE_CHECKSUM_BYTES
-    {
-        return Err(format!("update artifact checksum is too large: {len} bytes").into());
-    }
-
-    let response = response
-        .into_buffered_with_limit(
-            Some(MAX_UPDATE_CHECKSUM_BYTES),
-            None,
-            "update artifact checksum exceeded maximum allowed size",
-        )
-        .await?;
-
-    if response.body().len() > MAX_UPDATE_CHECKSUM_BYTES as usize {
-        return Err("update artifact checksum exceeded maximum allowed size".into());
-    }
-
-    let checksum = String::from_utf8(response.body().to_vec()).map_err(|_| {
-        FetchError::Message("update artifact checksum is not valid UTF-8".to_string())
-    })?;
-    parse_sha256_checksum(&checksum)
-}
-
-fn parse_sha256_checksum(contents: &str) -> Result<String, FetchError> {
-    let digest: String = contents.trim_start().chars().take(64).collect();
-    if digest.len() != 64 || !digest.bytes().all(|byte| byte.is_ascii_hexdigit()) {
-        return Err("update artifact checksum does not start with a SHA-256 digest".into());
-    }
-    Ok(digest.to_ascii_lowercase())
-}
-
-fn verify_artifact_checksum(
-    artifact_name: &str,
-    actual: &str,
-    expected: &str,
-) -> Result<(), FetchError> {
-    if actual != expected {
-        return Err(FetchError::Message(format!(
-            "update artifact checksum mismatch for {artifact_name}: expected {expected}, got {actual}"
-        )));
-    }
-    Ok(())
-}
-
-#[cfg(test)]
-fn sha256_hex(bytes: &[u8]) -> String {
-    let mut hasher = Sha256::new();
-    hasher.update(bytes);
-    finish_sha256_hex(hasher)
-}
-
-fn finish_sha256_hex(hasher: Sha256) -> String {
-    let digest = hasher.finalize();
-    hex_encode(&digest)
-}
-
-fn hex_encode(bytes: &[u8]) -> String {
-    const HEX: &[u8; 16] = b"0123456789abcdef";
-    let mut out = String::with_capacity(bytes.len() * 2);
-    for &byte in bytes {
-        out.push(HEX[(byte >> 4) as usize] as char);
-        out.push(HEX[(byte & 0x0f) as usize] as char);
-    }
-    out
-}
-
-async fn update_get(client: &UpdateClient<'_>, url: &str) -> Result<UpdateResponse, FetchError> {
-    client.get(url).await
-}
-
-async fn update_get_stream(
-    client: &UpdateClient<'_>,
-    url: &str,
-) -> Result<UpdateStreamingResponse, FetchError> {
-    client.get_stream(url).await
-}
-
-impl<'a> UpdateClient<'a> {
-    fn new(cli: &'a Cli) -> Result<Self, FetchError> {
-        let timeout = cli
-            .timeout
-            .map(|seconds| duration_from_seconds("timeout", seconds))
-            .transpose()?;
-        let connect_timeout = cli
-            .connect_timeout
-            .map(|seconds| duration_from_seconds("connect-timeout", seconds))
-            .transpose()?;
-        Ok(Self {
-            cli: Some(cli),
-            timeout,
-            connect_timeout,
-            allow_insecure_http: internal_update_url_override_allows_insecure_http(),
-            client: Mutex::new(None),
-        })
-    }
-
-    #[cfg(test)]
-    fn test(timeout: Option<Duration>) -> Self {
-        Self {
-            cli: None,
-            timeout,
-            connect_timeout: None,
-            allow_insecure_http: false,
-            client: Mutex::new(None),
-        }
-    }
-
-    #[cfg(test)]
-    fn test_allow_insecure_http(timeout: Option<Duration>) -> Self {
-        Self {
-            cli: None,
-            timeout,
-            connect_timeout: None,
-            allow_insecure_http: true,
-            client: Mutex::new(None),
-        }
-    }
-
-    #[cfg(test)]
-    fn test_with_cli_allow_insecure_http(cli: &'a Cli, timeout: Option<Duration>) -> Self {
-        Self {
-            cli: Some(cli),
-            timeout,
-            connect_timeout: None,
-            allow_insecure_http: true,
-            client: Mutex::new(None),
-        }
-    }
-
-    async fn get(&self, raw_url: &str) -> Result<UpdateResponse, FetchError> {
-        self.get_stream(raw_url).await?.into_buffered().await
-    }
-
-    async fn get_stream(&self, raw_url: &str) -> Result<UpdateStreamingResponse, FetchError> {
-        let mut url = url::Url::parse(raw_url)?;
-        validate_update_url(&url, self.allow_insecure_http)?;
-        let request_start = Instant::now();
-        let budget = TimeoutBudget::started_at(self.timeout, request_start);
-        let mut redirects = 0usize;
-        loop {
-            let client = budget
-                .run(Box::pin(async { self.client_for_url(&url).await }))
-                .await?;
-            let headers = update_headers();
-            let mut request = client
-                .client
-                .request(Method::GET, url.clone())
-                .headers(headers);
-            if let Some(timeout) = budget.remaining()? {
-                request = request.timeout(timeout);
-            }
-            let response = budget
-                .run(Box::pin(async move {
-                    request.send().await.map_err(FetchError::from)
-                }))
-                .await?;
-            if !is_update_redirect(response.status()) {
-                return Ok(UpdateStreamingResponse { response, budget });
-            }
-            let location = response
-                .headers()
-                .get(LOCATION)
-                .and_then(|value| value.to_str().ok())
-                .ok_or_else(|| {
-                    FetchError::Runtime("redirect response missing Location".to_string())
-                })?;
-            if redirects >= 10 {
-                return Err(FetchError::Runtime(
-                    "exceeded maximum number of redirects: 10".to_string(),
-                ));
-            }
-            url = update_redirect_target(&url, location, self.allow_insecure_http)?;
-            redirects += 1;
-        }
-    }
-
-    async fn client_for_url(&self, url: &url::Url) -> Result<client::UrlClient, FetchError> {
-        let origin = update_client_origin(url)?;
-        {
-            let cache = self.client.lock().await;
-            if let Some(cached) = cache.as_ref().filter(|cached| cached.origin == origin) {
-                return Ok(cached.client.clone());
-            }
-        }
-
-        let client = self.build_client_for_url(url).await?;
-        let mut cache = self.client.lock().await;
-        if let Some(cached) = cache.as_ref().filter(|cached| cached.origin == origin) {
-            return Ok(cached.client.clone());
-        }
-        *cache = Some(UpdateCachedClient {
-            origin,
-            client: client.clone(),
-        });
-        Ok(client)
-    }
-
-    async fn build_client_for_url(&self, url: &url::Url) -> Result<client::UrlClient, FetchError> {
-        let Some(cli) = self.cli else {
-            let mut builder = transport::Client::builder()
-                .use_rustls_tls()
-                .no_brotli()
-                .no_gzip()
-                .no_zstd();
-            if let Some(timeout) = self.connect_timeout {
-                builder = builder.connect_timeout(timeout);
-            }
-            return Ok(client::UrlClient {
-                client: builder.build()?,
-                dns_resolution: None,
-            });
-        };
-
-        let context = client::ClientBuildContext {
-            mode: client::ClientMode::Request(None),
-            request_timeout: None,
-            connect_timeout: self.connect_timeout,
-            request_start: Instant::now(),
-            session: None,
-            connect_timing: None,
-        };
-        client::build_client_for_url(cli, url, &context).await
-    }
-}
-
-fn validate_update_url(url: &url::Url, allow_insecure_http: bool) -> Result<(), FetchError> {
-    match url.scheme() {
-        "https" => Ok(()),
-        "http" if allow_insecure_http => Ok(()),
-        "http" => Err(FetchError::Message(
-            "refusing insecure self-update URL: self-update downloads require HTTPS".to_string(),
-        )),
-        scheme => Err(FetchError::Message(format!(
-            "unsupported self-update URL scheme '{scheme}': self-update downloads require HTTPS"
-        ))),
-    }
-}
-
-fn update_redirect_target(
-    current_url: &url::Url,
-    location: &str,
-    allow_insecure_http: bool,
-) -> Result<url::Url, FetchError> {
-    let url = current_url
-        .join(location)
-        .map_err(|err| FetchError::Runtime(format!("invalid redirect location: {err}")))?;
-    validate_update_url(&url, allow_insecure_http)?;
-    Ok(url)
-}
-
-fn update_client_origin(url: &url::Url) -> Result<String, FetchError> {
-    let authority = crate::net::url_authority(url)?;
-    Ok(format!("{}://{authority}", url.scheme()))
-}
-
-fn update_headers() -> HeaderMap {
-    let mut headers = HeaderMap::new();
-    headers.insert(
-        USER_AGENT,
-        HeaderValue::from_str(&core::user_agent()).expect("valid user agent"),
-    );
-    headers.insert(
-        ACCEPT,
-        HeaderValue::from_static(core::DEFAULT_ACCEPT_HEADER),
-    );
-    headers
-}
-
-fn is_update_redirect(status: StatusCode) -> bool {
-    matches!(
-        status,
-        StatusCode::MOVED_PERMANENTLY
-            | StatusCode::FOUND
-            | StatusCode::SEE_OTHER
-            | StatusCode::TEMPORARY_REDIRECT
-            | StatusCode::PERMANENT_REDIRECT
-    )
-}
-
-#[derive(Debug)]
-struct UpdateResponse {
-    status: StatusCode,
-    body: bytes::Bytes,
-}
-
-impl UpdateResponse {
-    fn status(&self) -> StatusCode {
-        self.status
-    }
-
-    fn body(&self) -> &[u8] {
-        &self.body
-    }
-}
-
-struct UpdateStreamingResponse {
-    response: transport::Response,
-    budget: TimeoutBudget,
-}
-
-impl UpdateStreamingResponse {
-    fn status(&self) -> StatusCode {
-        self.response.status()
-    }
-
-    fn headers(&self) -> &HeaderMap {
-        self.response.headers()
-    }
-
-    fn content_length(&self) -> Option<u64> {
-        content_length(self.headers())
-    }
-
-    async fn into_buffered(self) -> Result<UpdateResponse, FetchError> {
-        self.into_buffered_with_limit(None, None, "response body exceeded maximum allowed size")
-            .await
-    }
-
-    async fn into_buffered_with_limit(
-        self,
-        max_body_bytes: Option<u64>,
-        mut progress: Option<&mut UpdateDownloadProgress>,
-        limit_error: &'static str,
-    ) -> Result<UpdateResponse, FetchError> {
-        let budget = self.budget;
-        let status = self.response.status();
-        let headers = self.response.headers().clone();
-        let (mut body, _) = self.response.into_body_with_deadline();
-        let capacity = content_length(&headers)
-            .map(|len| max_body_bytes.map_or(len, |max| len.min(max)))
-            .and_then(|len| usize::try_from(len).ok())
-            .unwrap_or(0);
-        let mut bytes = Vec::with_capacity(capacity);
-        while let Some(frame) = budget
-            .run(Box::pin(async {
-                match body.frame().await {
-                    Some(Ok(frame)) => Ok(Some(frame)),
-                    Some(Err(err)) => {
-                        Err(FetchError::Runtime(format!("response body error: {err}")))
-                    }
-                    None => Ok(None),
-                }
-            }))
-            .await?
-        {
-            let Ok(data) = frame.into_data() else {
-                continue;
-            };
-            if data.is_empty() {
-                continue;
-            }
-            if let Some(max) = max_body_bytes {
-                let current = u64::try_from(bytes.len()).unwrap_or(u64::MAX);
-                let chunk = u64::try_from(data.len()).unwrap_or(u64::MAX);
-                if current.saturating_add(chunk) > max {
-                    return Err(limit_error.into());
-                }
-            }
-            if let Some(progress) = progress.as_deref_mut() {
-                progress.add(data.len());
-            }
-            bytes.extend_from_slice(&data);
-        }
-        Ok(UpdateResponse {
-            status,
-            body: bytes::Bytes::from(bytes),
-        })
-    }
-}
-
-fn content_length(headers: &HeaderMap) -> Option<u64> {
-    headers
-        .get(CONTENT_LENGTH)
-        .and_then(|value| value.to_str().ok())
-        .and_then(|value| value.parse().ok())
-}
-
-struct ChannelReader {
-    rx: mpsc::Receiver<bytes::Bytes>,
-    current: Option<bytes::Bytes>,
-    offset: usize,
-}
-
-impl ChannelReader {
-    fn new(rx: mpsc::Receiver<bytes::Bytes>) -> Self {
-        Self {
-            rx,
-            current: None,
-            offset: 0,
-        }
-    }
-}
-
-impl std::io::Read for ChannelReader {
-    fn read(&mut self, out: &mut [u8]) -> std::io::Result<usize> {
-        if out.is_empty() {
-            return Ok(0);
-        }
-
-        loop {
-            if let Some(current) = &self.current {
-                if self.offset < current.len() {
-                    let remaining = &current[self.offset..];
-                    let len = remaining.len().min(out.len());
-                    out[..len].copy_from_slice(&remaining[..len]);
-                    self.offset += len;
-                    if self.offset == current.len() {
-                        self.current = None;
-                        self.offset = 0;
-                    }
-                    return Ok(len);
-                }
-                self.current = None;
-                self.offset = 0;
-            }
-
-            match self.rx.blocking_recv() {
-                Some(bytes) if bytes.is_empty() => {}
-                Some(bytes) => self.current = Some(bytes),
-                None => return Ok(0),
-            }
-        }
-    }
-}
-
-enum UpdateDownloadProgress {
-    Bar {
-        counter: BarCounter,
-        printer: ProgressPrinter,
-    },
-    Spinner {
-        counter: SpinnerCounter,
-        printer: ProgressPrinter,
-    },
-    None,
-}
-
-impl UpdateDownloadProgress {
-    fn maybe_start(silent: bool, content_length: i64) -> Self {
-        if silent || !core::stdio().stderr_is_terminal() {
-            return Self::None;
-        }
-        Self::new_for_terminal(ProgressPrinter::stderr_with_color(false), content_length)
-    }
-
-    fn new_for_terminal(printer: ProgressPrinter, content_length: i64) -> Self {
-        if content_length > 0 {
-            Self::Bar {
-                counter: BarCounter::new(printer.clone(), content_length),
-                printer,
-            }
-        } else {
-            Self::Spinner {
-                counter: SpinnerCounter::new(printer.clone()),
-                printer,
-            }
-        }
-    }
-
-    fn add(&self, bytes: usize) {
-        match self {
-            Self::Bar { counter, .. } => counter.add(bytes as i64),
-            Self::Spinner { counter, .. } => counter.add(bytes as i64),
-            Self::None => {}
-        }
-    }
-
-    fn finish(&mut self) {
-        match self {
-            Self::Bar { counter, printer } => {
-                counter.stop();
-                progress::clear_line(printer, 60);
-            }
-            Self::Spinner { counter, printer } => {
-                counter.stop();
-                progress::clear_line(printer, 40);
-            }
-            Self::None => {}
-        }
-    }
-}
-
-fn release_artifact(release: &Release) -> Option<ReleaseArtifact<'_>> {
-    release_artifact_for_platform(release, goos(), goarch())
-}
-
-fn release_artifact_for_platform<'a>(
-    release: &'a Release,
-    goos: &str,
-    goarch: &str,
-) -> Option<ReleaseArtifact<'a>> {
-    let want = format!(
-        "fetch-{}-{}-{}.{}",
-        release.tag_name,
-        goos,
-        goarch,
-        artifact_suffix_for_goos(goos)
-    );
-    let archive = release.assets.iter().find(|asset| asset.name == want)?;
-    let checksum_name = format!("{want}.sha256");
-    let checksum = release
-        .assets
-        .iter()
-        .find(|asset| asset.name == checksum_name)?;
-    Some(ReleaseArtifact {
-        archive_name: archive.name.as_str(),
-        archive_url: archive.browser_download_url.as_str(),
-        checksum_url: checksum.browser_download_url.as_str(),
-    })
-}
-
-#[cfg(test)]
-fn unpack_artifact(dir: &Path, archive_name: &str, data: &[u8]) -> Result<(), FetchError> {
-    if archive_name.ends_with(".zip") {
-        unpack_zip_artifact(dir, data)
-    } else if archive_name.ends_with(".tar.gz") || archive_name.ends_with(".tgz") {
-        unpack_tar_gz_artifact(dir, data)
-    } else {
-        Err(format!("unsupported self-update archive format: {archive_name}").into())
-    }
-}
-
-struct UpdateTempDir {
-    path: PathBuf,
-}
-
-impl UpdateTempDir {
-    fn path(&self) -> &Path {
-        &self.path
-    }
-}
-
-impl Drop for UpdateTempDir {
-    fn drop(&mut self) {
-        let _ = std::fs::remove_dir_all(&self.path);
-    }
-}
-
-fn create_update_temp_dir() -> Result<UpdateTempDir, FetchError> {
-    let base = std::env::temp_dir();
-    for _ in 0..100 {
-        let path = base.join(format!("fetch-update-{}", unique_suffix()));
-        match create_private_dir(&path) {
-            Ok(()) => return Ok(UpdateTempDir { path }),
-            Err(err) if err.kind() == std::io::ErrorKind::AlreadyExists => continue,
-            Err(err) => return Err(err.into()),
-        }
-    }
-
-    Err(FetchError::Message(
-        "unable to create unique temporary update directory".to_string(),
-    ))
-}
-
-#[cfg(unix)]
-fn create_private_dir(path: &Path) -> std::io::Result<()> {
-    use std::os::unix::fs::DirBuilderExt;
-
-    std::fs::DirBuilder::new().mode(0o700).create(path)
-}
-
-#[cfg(not(unix))]
-fn create_private_dir(path: &Path) -> std::io::Result<()> {
-    std::fs::create_dir(path)
-}
-
-#[cfg(test)]
-fn unpack_tar_gz_artifact(dir: &Path, data: &[u8]) -> Result<(), FetchError> {
-    unpack_tar_gz_artifact_with_limits(dir, data, ArchiveExtractionLimits::default())
-}
-
-#[cfg(test)]
-fn unpack_tar_gz_artifact_with_limits(
-    dir: &Path,
-    data: &[u8],
-    limits: ArchiveExtractionLimits,
-) -> Result<(), FetchError> {
-    unpack_tar_gz_artifact_from_reader(dir, data, limits)
-}
-
-fn unpack_tar_gz_artifact_from_reader<R: std::io::Read>(
-    dir: &Path,
-    reader: R,
-    limits: ArchiveExtractionLimits,
-) -> Result<(), FetchError> {
-    let decoder = GzDecoder::new(reader);
-    let mut archive = tar::Archive::new(decoder);
-    let mut state = ArchiveExtractionState::new(limits);
-    for entry in archive.entries()? {
-        let mut entry = entry?;
-        state.account_entry()?;
-        let path = entry.path()?.to_string_lossy().into_owned();
-        let out = dir.join(safe_archive_path(&path)?);
-        if let Some(parent) = out.parent() {
-            std::fs::create_dir_all(parent)?;
-        }
-        if entry.header().entry_type().is_dir() {
-            std::fs::create_dir_all(&out)?;
-            continue;
-        }
-        if !entry.header().entry_type().is_file() {
-            continue;
-        }
-        copy_archive_entry_to_file(&mut entry, &out, &mut state)?;
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-
-            let mode = entry.header().mode().unwrap_or(0o755) & 0o777;
-            std::fs::set_permissions(&out, std::fs::Permissions::from_mode(mode))?;
-        }
-    }
-    Ok(())
-}
-
-#[cfg(test)]
-fn unpack_zip_artifact(dir: &Path, data: &[u8]) -> Result<(), FetchError> {
-    unpack_zip_artifact_with_limits(dir, data, ArchiveExtractionLimits::default())
-}
-
-#[cfg(test)]
-fn unpack_zip_artifact_with_limits(
-    dir: &Path,
-    data: &[u8],
-    limits: ArchiveExtractionLimits,
-) -> Result<(), FetchError> {
-    let reader = std::io::Cursor::new(data);
-    unpack_zip_artifact_from_reader(dir, reader, limits)
-}
-
-fn unpack_zip_artifact_from_file(dir: &Path, path: &Path) -> Result<(), FetchError> {
-    let file = std::fs::File::open(path)?;
-    unpack_zip_artifact_from_reader(dir, file, ArchiveExtractionLimits::default())
-}
-
-fn unpack_zip_artifact_from_reader<R: std::io::Read + std::io::Seek>(
-    dir: &Path,
-    reader: R,
-    limits: ArchiveExtractionLimits,
-) -> Result<(), FetchError> {
-    let mut archive =
-        zip::ZipArchive::new(reader).map_err(|err| FetchError::Message(format!("zip: {err}")))?;
-    let mut state = ArchiveExtractionState::new(limits);
-
-    for index in 0..archive.len() {
-        state.account_entry()?;
-        let mut file = archive
-            .by_index(index)
-            .map_err(|err| FetchError::Message(format!("zip: {err}")))?;
-        let out = dir.join(safe_archive_path(file.name())?);
-
-        if file.is_dir() {
-            std::fs::create_dir_all(&out)?;
-            continue;
-        }
-        if let Some(parent) = out.parent() {
-            std::fs::create_dir_all(parent)?;
-        }
-
-        copy_archive_entry_to_file(&mut file, &out, &mut state)?;
-
-        #[cfg(unix)]
-        if let Some(mode) = file.unix_mode() {
-            use std::os::unix::fs::PermissionsExt;
-            std::fs::set_permissions(&out, std::fs::Permissions::from_mode(mode & 0o777))?;
-        }
-    }
-
-    Ok(())
-}
-
-#[derive(Clone, Copy)]
-struct ArchiveExtractionLimits {
-    max_entries: usize,
-    max_uncompressed_bytes: u64,
-}
-
-impl Default for ArchiveExtractionLimits {
-    fn default() -> Self {
-        Self {
-            max_entries: MAX_UPDATE_ARCHIVE_ENTRIES,
-            max_uncompressed_bytes: MAX_UPDATE_UNPACKED_BYTES,
-        }
-    }
-}
-
-struct ArchiveExtractionState {
-    limits: ArchiveExtractionLimits,
-    entries: usize,
-    uncompressed_bytes: u64,
-}
-
-impl ArchiveExtractionState {
-    fn new(limits: ArchiveExtractionLimits) -> Self {
-        Self {
-            limits,
-            entries: 0,
-            uncompressed_bytes: 0,
-        }
-    }
-
-    fn account_entry(&mut self) -> Result<(), FetchError> {
-        self.entries = self.entries.checked_add(1).ok_or_else(|| {
-            FetchError::Message("self-update archive entry count overflowed".to_string())
-        })?;
-        if self.entries > self.limits.max_entries {
-            return Err(FetchError::Message(format!(
-                "self-update archive contains too many entries: maximum allowed is {}",
-                self.limits.max_entries
-            )));
-        }
-        Ok(())
-    }
-
-    fn account_bytes(&mut self, bytes: u64) -> Result<(), FetchError> {
-        self.uncompressed_bytes = self.uncompressed_bytes.saturating_add(bytes);
-        if self.uncompressed_bytes > self.limits.max_uncompressed_bytes {
-            return Err(FetchError::Message(format!(
-                "self-update archive uncompressed data exceeded maximum allowed size of {} bytes",
-                self.limits.max_uncompressed_bytes
-            )));
-        }
-        Ok(())
-    }
-
-    fn remaining_bytes(&self) -> u64 {
-        self.limits
-            .max_uncompressed_bytes
-            .saturating_sub(self.uncompressed_bytes)
-    }
-}
-
-fn copy_archive_entry_to_file<R: std::io::Read>(
-    reader: &mut R,
-    out: &Path,
-    state: &mut ArchiveExtractionState,
-) -> Result<u64, FetchError> {
-    let mut file = std::fs::File::create(out)?;
-    match copy_archive_entry_bounded(reader, &mut file, state) {
-        Ok(bytes) => Ok(bytes),
-        Err(err) => {
-            drop(file);
-            let _ = std::fs::remove_file(out);
-            Err(err)
-        }
-    }
-}
-
-fn copy_archive_entry_bounded<R: std::io::Read, W: std::io::Write>(
-    reader: &mut R,
-    writer: &mut W,
-    state: &mut ArchiveExtractionState,
-) -> Result<u64, FetchError> {
-    let remaining = state.remaining_bytes();
-    let mut limited = std::io::Read::take(&mut *reader, remaining.saturating_add(1));
-    let copied = std::io::copy(&mut limited, writer)?;
-    state.account_bytes(copied)?;
-    Ok(copied)
-}
-
-fn safe_archive_path(name: &str) -> Result<PathBuf, FetchError> {
-    if name.is_empty()
-        || name.starts_with('/')
-        || name.starts_with('\\')
-        || has_windows_drive_prefix(name)
-    {
-        return Err(format!("refusing to unpack unsafe path '{name}'").into());
-    }
-
-    let mut out = PathBuf::new();
-    for component in name.split(['/', '\\']) {
-        match component {
-            "" | "." => {}
-            ".." => return Err(format!("refusing to unpack unsafe path '{name}'").into()),
-            value if has_unsafe_windows_archive_component(value) => {
-                return Err(format!("refusing to unpack unsafe path '{name}'").into());
-            }
-            value => out.push(value),
-        }
-    }
-
-    if out.as_os_str().is_empty() {
-        return Err(format!("refusing to unpack unsafe path '{name}'").into());
-    }
-    Ok(out)
-}
-
-fn has_windows_drive_prefix(name: &str) -> bool {
-    let bytes = name.as_bytes();
-    bytes.len() >= 2 && bytes[1] == b':' && bytes[0].is_ascii_alphabetic()
-}
-
-fn has_unsafe_windows_archive_component(component: &str) -> bool {
-    if component.contains(':') {
-        return true;
-    }
-
-    let normalized = component.trim_end_matches([' ', '.']);
-    if normalized.len() != component.len() {
-        return true;
-    }
-    let raw_basename = normalized
-        .split_once('.')
-        .map_or(normalized, |(base, _)| base);
-    let basename = raw_basename.trim_end_matches([' ', '.']);
-    if basename.len() != raw_basename.len() {
-        return true;
-    }
-    if basename.is_empty() {
-        return true;
-    }
-    let uppercase = basename.to_ascii_uppercase();
-    matches!(uppercase.as_str(), "CON" | "PRN" | "AUX" | "NUL")
-        || matches!(
-            uppercase.as_bytes(),
-            [b'C', b'O', b'M', b'1'..=b'9'] | [b'L', b'P', b'T', b'1'..=b'9']
-        )
-}
-
-#[cfg(windows)]
-fn self_replace(exe_path: &Path, new_exe_path: &Path) -> Result<(), FetchError> {
-    let dir = exe_path.parent().unwrap_or_else(|| Path::new("."));
-    let temp_exe_path = create_temp_file_path(dir, TEMP_EXE_SUFFIX);
-    copy_file(&temp_exe_path, new_exe_path)?;
-
-    let old_exe_path = create_temp_file_path(dir, RELOCATED_SUFFIX);
-    std::fs::rename(exe_path, &old_exe_path).map_err(|err| {
-        let _ = std::fs::remove_file(&temp_exe_path);
-        FetchError::from(err)
-    })?;
-
-    if let Err(err) = std::fs::rename(&temp_exe_path, exe_path) {
-        return match std::fs::rename(&old_exe_path, exe_path) {
-            Ok(()) => Err(err.into()),
-            Err(rollback_err) => Err(FetchError::Message(format!(
-                "{err}; rollback failed: {rollback_err}"
-            ))),
-        };
-    }
-
-    schedule_self_deletion_on_shutdown(&old_exe_path)
-}
-
-#[cfg(not(windows))]
-fn self_replace(exe_path: &Path, new_exe_path: &Path) -> Result<(), FetchError> {
-    let dir = exe_path.parent().unwrap_or_else(|| Path::new("."));
-    let temp_path = create_temp_file_path(dir, ".__temp");
-    if let Err(err) = copy_file(&temp_path, new_exe_path) {
-        let _ = std::fs::remove_file(&temp_path);
-        return Err(err);
-    }
-
-    if let Err(err) = crate::fileutil::atomic_replace_file(&temp_path, exe_path) {
-        let _ = std::fs::remove_file(&temp_path);
-        return Err(err.into());
-    }
-
-    Ok(())
-}
-
-fn create_temp_file_path(dir: &Path, suffix: &str) -> PathBuf {
-    dir.join(format!(".fetch.{}{suffix}", unique_suffix()))
-}
-
-fn copy_file(dst: &Path, src: &Path) -> Result<(), FetchError> {
-    let metadata = std::fs::metadata(src)?;
-    std::fs::copy(src, dst)?;
-    std::fs::set_permissions(dst, metadata.permissions())?;
-    let file = std::fs::OpenOptions::new()
-        .read(true)
-        .write(true)
-        .open(dst)?;
-    file.sync_all()?;
-    Ok(())
-}
-
-#[cfg(windows)]
-fn schedule_self_deletion_on_shutdown(exe_path: &Path) -> Result<(), FetchError> {
-    let mut delete_path = exe_path.to_path_buf();
-    let mut exe_dir = exe_path
-        .parent()
-        .unwrap_or_else(|| Path::new("."))
-        .to_path_buf();
-    let relocated_exe_path = create_temp_file_path(&std::env::temp_dir(), RELOCATED_SUFFIX);
-    if std::fs::rename(exe_path, &relocated_exe_path).is_ok() {
-        delete_path = relocated_exe_path;
-        exe_dir = std::env::temp_dir();
-    }
-
-    let self_delete_path = create_temp_file_path(&exe_dir, SELF_DELETE_SUFFIX);
-    copy_file(&self_delete_path, &delete_path)?;
-
-    let self_delete_handle = open_delete_on_close_handle(&self_delete_path)?;
-    let duplicated_parent = duplicate_current_process_handle()?;
-    let spawn_result = spawn_self_delete_helper(&self_delete_path, duplicated_parent, &delete_path);
-    std::thread::sleep(Duration::from_millis(100));
-
-    // SAFETY: handles were returned by successful Win32 handle-opening APIs.
-    unsafe {
-        use windows_sys::Win32::Foundation::CloseHandle;
-        CloseHandle(duplicated_parent);
-        CloseHandle(self_delete_handle);
-    }
-
-    spawn_result
-}
-
-#[cfg(windows)]
-fn run_self_delete_helper(data: &str) -> Result<(), FetchError> {
-    use std::os::windows::process::CommandExt;
-    use windows_sys::Win32::Foundation::WAIT_OBJECT_0;
-    use windows_sys::Win32::Storage::FileSystem::DeleteFileW;
-    use windows_sys::Win32::System::Threading::{INFINITE, WaitForSingleObject};
-
-    let (parent_handle, original_path) = parse_self_delete_env_value(data)
-        .ok_or_else(|| FetchError::Message("invalid self-delete state".to_string()))?;
-    let parent_handle = parent_handle as windows_sys::Win32::Foundation::HANDLE;
-
-    // SAFETY: parent_handle is supplied by this executable's update launcher.
-    let wait_result = unsafe { WaitForSingleObject(parent_handle, INFINITE) };
-    if wait_result != WAIT_OBJECT_0 {
-        return Err("waiting for update parent process failed".into());
-    }
-
-    let original_wide = path_to_wide(&original_path);
-    // SAFETY: original_wide is null-terminated and points to valid memory.
-    let deleted = unsafe { DeleteFileW(original_wide.as_ptr()) };
-    if deleted == 0 {
-        return Err(std::io::Error::last_os_error().into());
-    }
-
-    let _ = Command::new("cmd.exe")
-        .args(["/c", "exit"])
-        .creation_flags(windows_sys::Win32::System::Threading::CREATE_NO_WINDOW)
-        .spawn();
-    Ok(())
-}
-
-#[cfg(windows)]
-fn open_delete_on_close_handle(
-    path: &Path,
-) -> Result<windows_sys::Win32::Foundation::HANDLE, FetchError> {
-    use windows_sys::Win32::Foundation::{GENERIC_READ, INVALID_HANDLE_VALUE};
-    use windows_sys::Win32::Security::SECURITY_ATTRIBUTES;
-    use windows_sys::Win32::Storage::FileSystem::{
-        CreateFileW, FILE_FLAG_DELETE_ON_CLOSE, FILE_SHARE_DELETE, FILE_SHARE_READ, OPEN_EXISTING,
-    };
-
-    let path_wide = path_to_wide(path);
-    let security = SECURITY_ATTRIBUTES {
-        nLength: std::mem::size_of::<SECURITY_ATTRIBUTES>() as u32,
-        lpSecurityDescriptor: std::ptr::null_mut(),
-        bInheritHandle: 1,
-    };
-
-    // SAFETY: path_wide is null-terminated and security points to a valid SECURITY_ATTRIBUTES.
-    let handle = unsafe {
-        CreateFileW(
-            path_wide.as_ptr(),
-            GENERIC_READ,
-            FILE_SHARE_READ | FILE_SHARE_DELETE,
-            &security,
-            OPEN_EXISTING,
-            FILE_FLAG_DELETE_ON_CLOSE,
-            std::ptr::null_mut(),
-        )
-    };
-    if handle == INVALID_HANDLE_VALUE {
-        Err(std::io::Error::last_os_error().into())
-    } else {
-        Ok(handle)
-    }
-}
-
-#[cfg(windows)]
-fn duplicate_current_process_handle() -> Result<windows_sys::Win32::Foundation::HANDLE, FetchError>
-{
-    use windows_sys::Win32::Foundation::{DUPLICATE_SAME_ACCESS, DuplicateHandle, HANDLE};
-    use windows_sys::Win32::System::Threading::GetCurrentProcess;
-
-    // SAFETY: GetCurrentProcess returns a pseudo-handle valid in this process.
-    let current = unsafe { GetCurrentProcess() };
-    let mut duplicated: HANDLE = std::ptr::null_mut();
-    // SAFETY: all handles refer to the current process and duplicated points to writable storage.
-    let ok = unsafe {
-        DuplicateHandle(
-            current,
-            current,
-            current,
-            &mut duplicated,
-            0,
-            1,
-            DUPLICATE_SAME_ACCESS,
-        )
-    };
-    if ok == 0 {
-        Err(std::io::Error::last_os_error().into())
-    } else {
-        Ok(duplicated)
-    }
-}
-
-#[cfg(windows)]
-fn spawn_self_delete_helper(
-    self_delete_path: &Path,
-    parent_handle: windows_sys::Win32::Foundation::HANDLE,
-    original_path: &Path,
-) -> Result<(), FetchError> {
-    use std::os::windows::process::CommandExt;
-    use windows_sys::Win32::System::Threading::CREATE_NO_WINDOW;
-
-    Command::new(self_delete_path)
-        .env(
-            SELF_DELETE_ENV,
-            self_delete_env_value(parent_handle as usize, original_path),
-        )
-        .creation_flags(CREATE_NO_WINDOW)
-        .spawn()?;
-    Ok(())
-}
-
-#[cfg(any(windows, test))]
-fn self_delete_env_value(parent_handle: usize, original_path: &Path) -> String {
-    format!("{parent_handle}_{}", original_path.display())
-}
-
-#[cfg(any(windows, test))]
-fn parse_self_delete_env_value(value: &str) -> Option<(usize, PathBuf)> {
-    let (handle, path) = value.split_once('_')?;
-    let handle = handle.parse::<usize>().ok()?;
-    Some((handle, PathBuf::from(path)))
-}
-
-#[cfg(any(windows, test))]
-fn path_has_suffix(path: &Path, suffix: &str) -> bool {
-    path.to_string_lossy().ends_with(suffix)
-}
-
-#[cfg(windows)]
-fn path_to_wide(path: &Path) -> Vec<u16> {
-    use std::os::windows::ffi::OsStrExt;
-
-    path.as_os_str().encode_wide().chain(Some(0)).collect()
-}
-
-fn write_update_success(silent: bool, old_version: &str, new_version: &str) {
-    if silent {
-        return;
-    }
-    eprintln!("Updated fetch: {old_version} -> {new_version}");
-
-    let compare_ref = changelog_compare_ref(old_version);
-    if !compare_ref.is_empty() {
-        eprintln!(
-            "\nChangelog: https://github.com/ryanfowler/fetch/compare/{compare_ref}...{new_version}"
-        );
-    }
-}
-
-fn write_msg(silent: bool, message: &str) {
-    if !silent {
-        eprint!("{message}");
-    }
-}
-
-fn current_exe() -> Result<PathBuf, FetchError> {
-    let path = std::env::current_exe()?;
-    std::fs::canonicalize(path).map_err(Into::into)
-}
-
-fn update_url() -> String {
-    internal_update_url_override().unwrap_or_else(|| "https://api.github.com".to_string())
-}
-
-fn internal_update_url_override() -> Option<String> {
-    std::env::var(INTERNAL_UPDATE_URL_ENV).ok()
-}
-
-fn internal_update_url_override_allows_insecure_http() -> bool {
-    internal_update_url_override()
-        .is_some_and(|raw_url| url::Url::parse(&raw_url).is_ok_and(|url| url.scheme() == "http"))
-}
-
-fn fetch_filename() -> &'static str {
-    if cfg!(windows) { "fetch.exe" } else { "fetch" }
-}
-
-fn artifact_suffix_for_goos(goos: &str) -> &'static str {
-    match goos {
-        "windows" => "zip",
-        _ => "tar.gz",
-    }
-}
-
-fn goos() -> &'static str {
-    if cfg!(target_os = "macos") {
-        "darwin"
-    } else if cfg!(target_os = "windows") {
-        "windows"
-    } else if cfg!(target_os = "linux") {
-        "linux"
-    } else if cfg!(target_os = "freebsd") {
-        "freebsd"
-    } else if cfg!(target_os = "openbsd") {
-        "openbsd"
-    } else if cfg!(target_os = "netbsd") {
-        "netbsd"
-    } else {
-        std::env::consts::OS
-    }
-}
-
-fn goarch() -> &'static str {
-    match std::env::consts::ARCH {
-        "x86_64" => "amd64",
-        "x86" => "386",
-        "aarch64" => "arm64",
-        "arm" => "arm",
-        "riscv64" => "riscv64",
-        "powerpc64" => "ppc64",
-        "powerpc64le" => "ppc64le",
-        "s390x" => "s390x",
-        arch => arch,
-    }
-}
-
-fn unique_suffix() -> String {
-    let nanos = SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .map(|duration| duration.as_nanos())
-        .unwrap_or_default();
-    format!("{}-{nanos}", std::process::id())
-}
-
-fn parse_auto_update_interval(value: &str) -> Option<Duration> {
-    let value = value.trim();
-    if value.is_empty() {
-        return None;
-    }
-
-    match value {
-        "false" | "FALSE" | "False" | "f" | "F" | "0" | "off" | "no" | "never" => return None,
-        "true" | "TRUE" | "True" | "t" | "T" | "1" | "on" | "yes" => {
-            return Some(Duration::from_secs(24 * 60 * 60));
-        }
-        "0s" | "0sec" | "0secs" | "0m" => return Some(Duration::ZERO),
-        _ => {}
-    }
-
-    crate::duration::parse_duration_interval(value)
-}
-
-fn changelog_compare_ref(old_version: &str) -> String {
-    if is_version_tag(old_version) {
-        old_version.to_string()
-    } else {
-        vcs_revision().unwrap_or_default()
-    }
-}
-
-fn is_version_tag(value: &str) -> bool {
-    if value.len() < 6 || !value.starts_with('v') {
-        return false;
-    }
-
-    let mut dots = 0;
-    let bytes = value.as_bytes();
-    for i in 1..bytes.len() {
-        match bytes[i] {
-            b'.' => {
-                if i == 1 || i == bytes.len() - 1 || bytes[i - 1] == b'.' {
-                    return false;
-                }
-                dots += 1;
-            }
-            b'0'..=b'9' => {}
-            _ => return false,
-        }
-    }
-    dots == 2
-}
-
-fn vcs_revision() -> Option<String> {
-    option_env!("FETCH_VCS_REVISION").and_then(|value| {
-        if value.is_empty() {
-            None
-        } else {
-            Some(value.to_string())
-        }
-    })
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-struct Metadata {
-    last_attempt_at: String,
-}
-
-fn record_last_attempt_time(dir: &Path) {
-    let _ = update_last_attempt_time(dir, SystemTime::now());
-}
-
-fn should_attempt_update(interval: Duration) -> Result<bool, FetchError> {
-    let dir = cache_dir()?;
-    should_attempt_update_in(&dir, interval)
-}
-
-fn should_attempt_update_in(dir: &Path, interval: Duration) -> Result<bool, FetchError> {
-    let Some(_lock) = acquire_update_lock(dir, false, true, None)? else {
-        return Ok(false);
-    };
-    let path = dir.join("metadata.json");
-    let data = match std::fs::read(&path) {
-        Ok(data) => data,
-        Err(err) if err.kind() == std::io::ErrorKind::NotFound => return Ok(true),
-        Err(err) => return Err(err.into()),
-    };
-
-    let metadata: Metadata = match serde_json::from_slice(&data) {
-        Ok(metadata) => metadata,
-        Err(_) => return Ok(true),
-    };
-    let last_attempt_at = match OffsetDateTime::parse(&metadata.last_attempt_at, &Rfc3339) {
-        Ok(value) => value,
-        Err(_) => return Ok(true),
-    };
-    let now = OffsetDateTime::now_utc();
-    let elapsed = (now - last_attempt_at).try_into().unwrap_or(Duration::ZERO);
-    Ok(elapsed > interval)
-}
-
-fn cache_dir() -> Result<PathBuf, FetchError> {
-    let base = user_cache_dir().ok_or_else(|| {
-        FetchError::Message("unable to determine user cache directory".to_string())
-    })?;
-    let dir = base.join("fetch");
-    std::fs::create_dir_all(&dir)?;
-    Ok(dir)
-}
-
-fn user_cache_dir() -> Option<PathBuf> {
-    if cfg!(target_os = "windows") {
-        std::env::var_os("LOCALAPPDATA").map(PathBuf::from)
-    } else if cfg!(target_os = "macos") {
-        std::env::var_os("HOME").map(|home| PathBuf::from(home).join("Library").join("Caches"))
-    } else if let Some(cache) = std::env::var_os("XDG_CACHE_HOME") {
-        Some(PathBuf::from(cache))
-    } else {
-        std::env::var_os("HOME").map(|home| PathBuf::from(home).join(".cache"))
-    }
-}
-
-fn update_last_attempt_time(dir: &Path, now: SystemTime) -> Result<(), FetchError> {
-    let timestamp = OffsetDateTime::from(now)
-        .to_offset(time::UtcOffset::UTC)
-        .format(&Rfc3339)
-        .map_err(|err| FetchError::Message(err.to_string()))?;
-    let data = serde_json::to_vec(&Metadata {
-        last_attempt_at: timestamp,
-    })
-    .map_err(|err| FetchError::Message(err.to_string()))?;
-
-    std::fs::create_dir_all(dir)?;
-    let path = dir.join("metadata.json");
-    let tmp = dir.join(format!(".metadata-{}.tmp", unique_suffix()));
-    std::fs::write(&tmp, data)?;
-    if let Err(err) = crate::fileutil::atomic_replace_file(&tmp, &path) {
-        let _ = std::fs::remove_file(&tmp);
-        return Err(err.into());
-    }
-    Ok(())
-}
-
-fn acquire_update_lock(
-    dir: &Path,
-    block: bool,
-    silent: bool,
-    color: Option<&str>,
-) -> Result<Option<FileLock>, FetchError> {
-    acquire_update_lock_with_timeout(dir, block, silent, color, UPDATE_LOCK_WAIT_TIMEOUT)
-}
-
-fn acquire_update_lock_with_timeout(
-    dir: &Path,
-    block: bool,
-    silent: bool,
-    color: Option<&str>,
-    timeout: Duration,
-) -> Result<Option<FileLock>, FetchError> {
-    std::fs::create_dir_all(dir)?;
-    let path = dir.join(".update-lock");
-    if !block {
-        return FileLock::try_acquire(path).map_err(FetchError::from);
-    }
-
-    FileLock::acquire_with_timeout(
-        path,
-        timeout,
-        || {
-            if !silent {
-                write_warning_with_separator_with_color("waiting on lock to begin updating", color);
-            }
-        },
-        update_lock_timeout_error,
-    )
-    .map(Some)
-}
-
-fn update_lock_wait_timeout(request_timeout: Option<Duration>) -> Duration {
-    request_timeout
-        .map(|timeout| timeout.min(UPDATE_LOCK_WAIT_TIMEOUT))
-        .unwrap_or(UPDATE_LOCK_WAIT_TIMEOUT)
-}
-
-fn update_lock_timeout_error(timeout: Duration) -> FetchError {
-    FetchError::Runtime(format!(
-        "timed out waiting for update lock after {}",
-        format_go_duration(timeout)
-    ))
-}
-
-#[cfg(unix)]
-fn can_replace_file(path: &Path) -> bool {
-    let dir = path.parent().unwrap_or_else(|| Path::new("."));
-    let temp_path = dir.join(format!(".fetch-update-{}", unique_suffix()));
-    match std::fs::File::create(&temp_path) {
-        Ok(file) => {
-            drop(file);
-            std::fs::remove_file(temp_path).is_ok()
-        }
-        Err(_) => false,
-    }
-}
-
-#[cfg(not(unix))]
-fn can_replace_file(_path: &Path) -> bool {
-    true
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use flate2::Compression;
-    use flate2::write::GzEncoder;
-    use std::io::{BufRead, Write};
-    use std::net::TcpListener;
-    use std::sync::atomic::{AtomicUsize, Ordering};
-    use std::sync::{Arc, Mutex};
-    use std::task::Poll;
-    use std::thread::JoinHandle;
-    use tar::{Builder, Header};
-
-    #[derive(Clone)]
-    struct SharedBuffer(Arc<Mutex<Vec<u8>>>);
-
-    impl Write for SharedBuffer {
-        fn write(&mut self, buf: &[u8]) -> std::io::Result<usize> {
-            self.0.lock().unwrap().extend_from_slice(buf);
-            Ok(buf.len())
-        }
-
-        fn flush(&mut self) -> std::io::Result<()> {
-            Ok(())
-        }
-    }
-
-    fn memory_printer() -> (ProgressPrinter, Arc<Mutex<Vec<u8>>>) {
-        let buffer = Arc::new(Mutex::new(Vec::new()));
-        (
-            ProgressPrinter::new(SharedBuffer(buffer.clone()), false),
-            buffer,
-        )
-    }
-
-    fn start_artifact_response(
-        headers: Vec<(&'static str, String)>,
-        chunks: Vec<Vec<u8>>,
-    ) -> (String, JoinHandle<()>) {
-        let listener = TcpListener::bind("127.0.0.1:0").unwrap();
-        let url = format!("http://{}", listener.local_addr().unwrap());
-        let join = std::thread::spawn(move || {
-            let (mut stream, _) = listener.accept().unwrap();
-            let mut reader = std::io::BufReader::new(stream.try_clone().unwrap());
-            let mut line = String::new();
-            loop {
-                line.clear();
-                if reader.read_line(&mut line).unwrap() == 0 || line == "\r\n" {
-                    break;
-                }
-            }
-
-            write!(stream, "HTTP/1.1 200 OK\r\nConnection: close\r\n").unwrap();
-            for (name, value) in headers {
-                write!(stream, "{name}: {value}\r\n").unwrap();
-            }
-            write!(stream, "\r\n").unwrap();
-            for chunk in chunks {
-                stream.write_all(&chunk).unwrap();
-            }
-        });
-        (url, join)
-    }
-
-    fn read_request_line_and_headers(stream: &mut std::net::TcpStream) -> String {
-        let mut reader = std::io::BufReader::new(stream.try_clone().unwrap());
-        let mut request_line = String::new();
-        reader.read_line(&mut request_line).unwrap();
-        let mut line = String::new();
-        loop {
-            line.clear();
-            if reader.read_line(&mut line).unwrap() == 0 || line == "\r\n" {
-                break;
-            }
-        }
-        request_line
-    }
-
-    fn start_update_proxy(body: &'static str) -> (String, JoinHandle<String>) {
-        let listener = TcpListener::bind("127.0.0.1:0").unwrap();
-        let url = format!("http://{}", listener.local_addr().unwrap());
-        let join = std::thread::spawn(move || {
-            let (mut stream, _) = listener.accept().unwrap();
-            let request_line = read_request_line_and_headers(&mut stream);
-            write!(
-                stream,
-                "HTTP/1.1 200 OK\r\nContent-Length: {}\r\nConnection: close\r\n\r\n{}",
-                body.len(),
-                body
-            )
-            .unwrap();
-            request_line
-        });
-        (url, join)
-    }
-
-    fn start_slow_redirect_response(
-        first_delay: Duration,
-        final_delay: Duration,
-    ) -> (String, JoinHandle<()>) {
-        let listener = TcpListener::bind("127.0.0.1:0").unwrap();
-        let url = format!("http://{}", listener.local_addr().unwrap());
-        let join = std::thread::spawn(move || {
-            if let Ok((mut stream, _)) = listener.accept() {
-                let _ = read_request_line_and_headers(&mut stream);
-                std::thread::sleep(first_delay);
-                let _ = write!(
-                    stream,
-                    "HTTP/1.1 302 Found\r\nLocation: /final\r\nContent-Length: 0\r\nConnection: close\r\n\r\n"
-                );
-            }
-            if let Ok((mut stream, _)) = listener.accept() {
-                let _ = read_request_line_and_headers(&mut stream);
-                std::thread::sleep(final_delay);
-                let _ = write!(
-                    stream,
-                    "HTTP/1.1 200 OK\r\nContent-Length: 2\r\nConnection: close\r\n\r\nok"
-                );
-            }
-        });
-        (url, join)
-    }
-
-    #[test]
-    fn update_requests_use_go_client_headers() {
-        let headers = update_headers();
-
-        assert_eq!(
-            headers.get(USER_AGENT).and_then(|v| v.to_str().ok()),
-            Some(core::user_agent().as_str())
-        );
-        assert_eq!(
-            headers.get(ACCEPT).and_then(|v| v.to_str().ok()),
-            Some(core::DEFAULT_ACCEPT_HEADER)
-        );
-    }
-
-    #[test]
-    fn auto_update_command_args_preserve_explicit_config_only() {
-        let args = auto_update_command_args(Some(Path::new("custom config.ini")));
-        let args = args
-            .iter()
-            .map(|arg| arg.to_string_lossy().into_owned())
-            .collect::<Vec<_>>();
-
-        assert_eq!(
-            args,
-            vec![
-                "--config",
-                "custom config.ini",
-                "--update",
-                "--timeout=300",
-                "--silent",
-            ]
-        );
-    }
-
-    #[test]
-    fn update_client_origin_keys_default_ports_and_ipv6_hosts() {
-        let url = url::Url::parse("https://example.com/releases/latest").unwrap();
-        assert_eq!(
-            update_client_origin(&url).unwrap(),
-            "https://example.com:443"
-        );
-
-        let url = url::Url::parse("http://[::1]/artifact").unwrap();
-        assert_eq!(update_client_origin(&url).unwrap(), "http://[::1]:80");
-    }
-
-    #[test]
-    fn update_redirect_target_rejects_https_to_http_by_default() {
-        let url = url::Url::parse("https://updates.example/start").unwrap();
-
-        let err =
-            update_redirect_target(&url, "http://updates.example/artifact", false).unwrap_err();
-
-        assert!(
-            err.to_string()
-                .contains("self-update downloads require HTTPS"),
-            "{err}"
-        );
-    }
-
-    #[tokio::test]
-    async fn update_client_rejects_initial_http_by_default() {
-        let client = UpdateClient::test(None);
-
-        let err = update_get(&client, "http://127.0.0.1:1/artifact")
-            .await
-            .unwrap_err();
-
-        assert!(
-            err.to_string()
-                .contains("self-update downloads require HTTPS"),
-            "{err}"
-        );
-    }
-
-    #[tokio::test]
-    async fn update_client_internal_test_override_allows_local_http_fixture() {
-        let (url, join) = start_artifact_response(
-            vec![("Content-Length", "2".to_string())],
-            vec![b"ok".to_vec()],
-        );
-        let client = UpdateClient::test_allow_insecure_http(None);
-
-        let response = update_get(&client, &url).await.unwrap();
-        join.join().unwrap();
-
-        assert_eq!(response.status(), StatusCode::OK);
-        assert_eq!(response.body(), b"ok");
-    }
-
-    #[tokio::test]
-    async fn download_artifact_rejects_content_length_above_limit() {
-        let (url, join) =
-            start_artifact_response(vec![("Content-Length", "11".to_string())], Vec::new());
-        let client = UpdateClient::test_allow_insecure_http(None);
-        let dir = tempfile::tempdir().unwrap();
-
-        let err = download_and_unpack_artifact_with_limit(
-            &client,
-            "fetch-v1.2.3-linux-amd64.tar.gz",
-            &url,
-            dir.path(),
-            dir.path(),
-            true,
-            10,
-        )
-        .await
-        .unwrap_err();
-        join.join().unwrap();
-
-        assert!(
-            err.to_string()
-                .contains("update artifact is too large: 11 bytes")
-        );
-    }
-
-    #[tokio::test]
-    async fn download_artifact_rejects_stream_above_limit() {
-        let (url, join) = start_artifact_response(
-            Vec::new(),
-            vec![b"12345".to_vec(), b"67890".to_vec(), b"!".to_vec()],
-        );
-        let client = UpdateClient::test_allow_insecure_http(None);
-        let dir = tempfile::tempdir().unwrap();
-
-        let err = download_and_unpack_artifact_with_limit(
-            &client,
-            "fetch-v1.2.3-linux-amd64.tar.gz",
-            &url,
-            dir.path(),
-            dir.path(),
-            true,
-            10,
-        )
-        .await
-        .unwrap_err();
-        join.join().unwrap();
-
-        assert!(
-            err.to_string()
-                .contains("update artifact exceeded maximum allowed size")
-        );
-    }
-
-    #[cfg(unix)]
-    #[tokio::test]
-    async fn download_artifact_streams_tar_gz_to_unpack_dir_and_hashes() {
-        let archive = create_tar_gz(&[("fetch", b"content".as_slice(), 0o755, false)]);
-        let expected = sha256_hex(&archive);
-        let (url, join) = start_artifact_response(
-            Vec::new(),
-            archive.chunks(3).map(|chunk| chunk.to_vec()).collect(),
-        );
-        let client = UpdateClient::test_allow_insecure_http(None);
-        let dir = tempfile::tempdir().unwrap();
-
-        let actual = download_and_unpack_artifact_with_limit(
-            &client,
-            "fetch-v1.2.3-linux-amd64.tar.gz",
-            &url,
-            dir.path(),
-            dir.path(),
-            true,
-            MAX_UPDATE_ARTIFACT_BYTES,
-        )
-        .await
-        .unwrap();
-        join.join().unwrap();
-
-        assert_eq!(actual, expected);
-        assert_eq!(
-            std::fs::read_to_string(dir.path().join("fetch")).unwrap(),
-            "content"
-        );
-    }
-
-    #[cfg(unix)]
-    #[tokio::test]
-    async fn unpack_tar_gz_response_stops_downloading_after_unpack_fails() {
-        const TAIL_CHUNKS: usize = 64;
-        const TAIL_CHUNK_LEN: usize = 64 * 1024;
-
-        let archive = create_tar_gz(&[("../escape.txt", b"bad".as_slice(), 0o644, false)]);
-        let pulled_chunks = Arc::new(AtomicUsize::new(0));
-        let stream_pulled_chunks = Arc::clone(&pulled_chunks);
-        let mut first_chunk = Some(bytes::Bytes::from(archive));
-        let mut tail_remaining = TAIL_CHUNKS;
-        let stream = futures_util::stream::poll_fn(move |_| {
-            if let Some(chunk) = first_chunk.take() {
-                stream_pulled_chunks.fetch_add(1, Ordering::SeqCst);
-                return Poll::Ready(Some(Ok::<_, std::io::Error>(chunk)));
-            }
-            if tail_remaining == 0 {
-                return Poll::Ready(None);
-            }
-            tail_remaining -= 1;
-            stream_pulled_chunks.fetch_add(1, Ordering::SeqCst);
-            Poll::Ready(Some(Ok::<_, std::io::Error>(bytes::Bytes::from(vec![
-                b'x';
-                TAIL_CHUNK_LEN
-            ]))))
-        });
-        let response = UpdateStreamingResponse {
-            response: transport::Response::test(
-                StatusCode::OK,
-                HeaderMap::new(),
-                transport::Body::wrap_stream(stream),
-            ),
-            budget: TimeoutBudget::started_at(None, Instant::now()),
-        };
-        let dir = tempfile::tempdir().unwrap();
-        let mut progress = UpdateDownloadProgress::None;
-
-        let err = unpack_tar_gz_response(
-            response,
-            dir.path(),
-            &mut progress,
-            MAX_UPDATE_ARTIFACT_BYTES,
-        )
-        .await
-        .unwrap_err();
-
-        assert!(
-            err.to_string()
-                .contains("refusing to unpack unsafe path '../escape.txt'"),
-            "{err}"
-        );
-        let pulled = pulled_chunks.load(Ordering::SeqCst);
-        assert!(
-            pulled < TAIL_CHUNKS + 1,
-            "expected early unpack failure to stop the download, pulled {pulled} of {} chunks",
-            TAIL_CHUNKS + 1
-        );
-    }
-
-    #[tokio::test]
-    async fn download_artifact_streams_zip_to_temp_file_and_hashes() {
-        let archive = create_zip(&[("fetch.exe", b"content".as_slice(), false)]);
-        let expected = sha256_hex(&archive);
-        let (url, join) = start_artifact_response(
-            Vec::new(),
-            archive.chunks(3).map(|chunk| chunk.to_vec()).collect(),
-        );
-        let client = UpdateClient::test_allow_insecure_http(None);
-        let dir = tempfile::tempdir().unwrap();
-        let unpack_dir = dir.path().join("unpacked");
-        std::fs::create_dir(&unpack_dir).unwrap();
-
-        let actual = download_and_unpack_artifact_with_limit(
-            &client,
-            "fetch-v1.2.3-windows-amd64.zip",
-            &url,
-            &unpack_dir,
-            dir.path(),
-            true,
-            MAX_UPDATE_ARTIFACT_BYTES,
-        )
-        .await
-        .unwrap();
-        join.join().unwrap();
-
-        assert_eq!(actual, expected);
-        assert_eq!(
-            std::fs::read_to_string(unpack_dir.join("fetch.exe")).unwrap(),
-            "content"
-        );
-    }
-
-    #[tokio::test]
-    async fn update_client_uses_shared_proxy_config() {
-        let (proxy_url, join) = start_update_proxy("proxied update");
-        let cli =
-            <Cli as clap::Parser>::try_parse_from(["fetch", "--update", "--proxy", &proxy_url])
-                .unwrap();
-        let client = UpdateClient::test_with_cli_allow_insecure_http(&cli, None);
-
-        let response = update_get(&client, "http://updates.example/artifact")
-            .await
-            .unwrap();
-        let request_line = join.join().unwrap();
-
-        assert_eq!(response.status(), StatusCode::OK);
-        assert_eq!(response.body(), b"proxied update");
-        assert!(
-            request_line.starts_with("GET http://updates.example/artifact HTTP/1.1"),
-            "{request_line}"
-        );
-    }
-
-    #[tokio::test]
-    async fn update_redirects_share_one_request_timeout_budget() {
-        let (url, join) =
-            start_slow_redirect_response(Duration::from_millis(150), Duration::from_millis(150));
-        let client = UpdateClient::test_allow_insecure_http(Some(Duration::from_millis(220)));
-
-        let err = update_get(&client, &url).await.unwrap_err();
-        let _ = join.join();
-
-        assert!(
-            err.to_string().contains("request timed out after 220ms"),
-            "{err}"
-        );
-    }
-
-    #[test]
-    fn release_artifact_matches_go_platform_names() {
-        let archive_name = format!(
-            "fetch-v1.2.3-{}-{}.{}",
-            goos(),
-            goarch(),
-            artifact_suffix_for_goos(goos())
-        );
-        let release = Release {
-            tag_name: "v1.2.3".to_string(),
-            assets: vec![
-                Asset {
-                    name: archive_name.clone(),
-                    browser_download_url: "https://example.test/artifact".to_string(),
-                },
-                Asset {
-                    name: format!("{archive_name}.sha256"),
-                    browser_download_url: "https://example.test/artifact.sha256".to_string(),
-                },
-            ],
-        };
-
-        assert_eq!(
-            release_artifact(&release),
-            Some(ReleaseArtifact {
-                archive_name: archive_name.as_str(),
-                archive_url: "https://example.test/artifact",
-                checksum_url: "https://example.test/artifact.sha256",
-            })
-        );
-    }
-
-    #[test]
-    fn release_artifact_uses_zip_for_windows_release_assets() {
-        let release = Release {
-            tag_name: "v1.2.3".to_string(),
-            assets: vec![
-                Asset {
-                    name: "fetch-v1.2.3-linux-amd64.tar.gz".to_string(),
-                    browser_download_url: "https://example.test/linux".to_string(),
-                },
-                Asset {
-                    name: "fetch-v1.2.3-windows-amd64.zip".to_string(),
-                    browser_download_url: "https://example.test/windows".to_string(),
-                },
-                Asset {
-                    name: "fetch-v1.2.3-windows-amd64.zip.sha256".to_string(),
-                    browser_download_url: "https://example.test/windows.sha256".to_string(),
-                },
-            ],
-        };
-
-        assert_eq!(
-            release_artifact_for_platform(&release, "windows", "amd64"),
-            Some(ReleaseArtifact {
-                archive_name: "fetch-v1.2.3-windows-amd64.zip",
-                archive_url: "https://example.test/windows",
-                checksum_url: "https://example.test/windows.sha256",
-            })
-        );
-        assert_eq!(
-            release_artifact_for_platform(&release, "windows", "arm64"),
-            None
-        );
-    }
-
-    #[test]
-    fn release_artifact_requires_checksum_sidecar() {
-        let release = Release {
-            tag_name: "v1.2.3".to_string(),
-            assets: vec![Asset {
-                name: "fetch-v1.2.3-linux-amd64.tar.gz".to_string(),
-                browser_download_url: "https://example.test/linux".to_string(),
-            }],
-        };
-
-        assert_eq!(
-            release_artifact_for_platform(&release, "linux", "amd64"),
-            None
-        );
-    }
-
-    #[test]
-    fn update_artifact_checksum_accepts_valid_sidecar_digest() {
-        let artifact = b"verified release bytes";
-        let checksum = format!(
-            "{}  fetch-v1.2.3-linux-amd64.tar.gz\n",
-            sha256_hex(artifact)
-        );
-
-        let parsed = parse_sha256_checksum(&checksum).unwrap();
-        let actual = sha256_hex(artifact);
-
-        verify_artifact_checksum("fetch-v1.2.3-linux-amd64.tar.gz", &actual, &parsed).unwrap();
-    }
-
-    #[test]
-    fn update_artifact_checksum_rejects_wrong_digest() {
-        let artifact = b"tampered release bytes";
-        let checksum = "0000000000000000000000000000000000000000000000000000000000000000";
-        let actual = sha256_hex(artifact);
-
-        let err = verify_artifact_checksum("fetch-v1.2.3-linux-amd64.tar.gz", &actual, checksum)
-            .unwrap_err();
-
-        assert!(
-            err.to_string()
-                .contains("update artifact checksum mismatch")
-        );
-    }
-
-    #[cfg(not(windows))]
-    #[test]
-    fn non_windows_self_replace_keeps_unpacked_source() {
-        let exe_dir = tempfile::tempdir().unwrap();
-        let unpack_dir = tempfile::tempdir().unwrap();
-        let exe_path = exe_dir.path().join(fetch_filename());
-        let new_exe_path = unpack_dir.path().join(fetch_filename());
-        std::fs::write(&exe_path, b"old executable").unwrap();
-        std::fs::write(&new_exe_path, b"new executable").unwrap();
-
-        self_replace(&exe_path, &new_exe_path).unwrap();
-
-        assert_eq!(std::fs::read(&exe_path).unwrap(), b"new executable");
-        assert_eq!(std::fs::read(&new_exe_path).unwrap(), b"new executable");
-        assert!(std::fs::read_dir(exe_dir.path()).unwrap().all(|entry| {
-            !entry
-                .unwrap()
-                .file_name()
-                .to_string_lossy()
-                .starts_with(".fetch.")
-        }));
-    }
-
-    #[test]
-    fn windows_self_delete_env_value_round_trips_paths_with_underscores() {
-        let path = PathBuf::from(r"C:\Program Files\fetch_cli\fetch.__relocated.exe");
-        let value = self_delete_env_value(12345, &path);
-
-        let (handle, parsed_path) = parse_self_delete_env_value(&value).unwrap();
-
-        assert_eq!(handle, 12345);
-        assert_eq!(parsed_path, path);
-    }
-
-    #[test]
-    fn windows_self_delete_env_value_rejects_invalid_handle() {
-        assert!(parse_self_delete_env_value("abc_C:\\fetch.exe").is_none());
-        assert!(parse_self_delete_env_value("12345").is_none());
-    }
-
-    #[test]
-    fn windows_temp_paths_use_go_self_replace_suffixes() {
-        let dir = Path::new("C:\\fetch");
-
-        let temp = create_temp_file_path(dir, TEMP_EXE_SUFFIX);
-        let relocated = create_temp_file_path(dir, RELOCATED_SUFFIX);
-        let self_delete = create_temp_file_path(dir, SELF_DELETE_SUFFIX);
-
-        assert!(path_has_suffix(&temp, TEMP_EXE_SUFFIX));
-        assert!(path_has_suffix(&relocated, RELOCATED_SUFFIX));
-        assert!(path_has_suffix(&self_delete, SELF_DELETE_SUFFIX));
-        assert!(
-            temp.file_name()
-                .unwrap()
-                .to_string_lossy()
-                .starts_with(".fetch.")
-        );
-    }
-
-    #[test]
-    fn auto_update_zero_interval_is_due() {
-        assert!(parse_auto_update_interval("0s").is_some());
-        assert!(parse_auto_update_interval("0sec").is_some());
-        assert_eq!(parse_auto_update_interval("0"), None);
-    }
-
-    #[test]
-    fn test_is_version_tag() {
-        let tests = [
-            ("v1.2.3", true),
-            ("v0.0.0", true),
-            ("v10.20.30", true),
-            ("v100.200.300", true),
-            ("v(dev)", false),
-            ("v0.0.0-20231215164305-abcdef123456", false),
-            ("1.2.3", false),
-            ("v1.2", false),
-            ("v1.2.3.4", false),
-            ("v.1.2", false),
-            ("v1..2", false),
-            ("v1.2.", false),
-            ("", false),
-            ("v", false),
-            ("vx.y.z", false),
-            ("v1.2.3-rc1", false),
-            ("v1.2.3+meta", false),
-        ];
-
-        for (input, want) in tests {
-            assert_eq!(is_version_tag(input), want, "is_version_tag({input:?})");
-        }
-    }
-
-    #[test]
-    fn update_last_attempt_time_overwrites_existing_metadata() {
-        let dir = tempfile::tempdir().unwrap();
-
-        update_last_attempt_time(dir.path(), UNIX_EPOCH + Duration::from_secs(100)).unwrap();
-        update_last_attempt_time(dir.path(), UNIX_EPOCH + Duration::from_secs(200)).unwrap();
-
-        let data = std::fs::read_to_string(dir.path().join("metadata.json")).unwrap();
-        assert_eq!(data, r#"{"last_attempt_at":"1970-01-01T00:03:20Z"}"#);
-    }
-
-    #[test]
-    fn auto_update_interval_parses_go_bool_and_duration_values() {
-        assert_eq!(
-            parse_auto_update_interval("true"),
-            Some(Duration::from_secs(24 * 60 * 60))
-        );
-        assert_eq!(
-            parse_auto_update_interval("1h30m"),
-            Some(Duration::from_secs(90 * 60))
-        );
-        assert_eq!(
-            parse_auto_update_interval("1.5h"),
-            Some(Duration::from_secs(90 * 60))
-        );
-        assert_eq!(
-            parse_auto_update_interval("+30m"),
-            Some(Duration::from_secs(30 * 60))
-        );
-        assert_eq!(
-            parse_auto_update_interval("1d"),
-            Some(Duration::from_secs(24 * 60 * 60))
-        );
-        assert_eq!(
-            parse_auto_update_interval("250ms"),
-            Some(Duration::from_millis(250))
-        );
-        assert_eq!(parse_auto_update_interval("false"), None);
-        assert_eq!(parse_auto_update_interval("-1h"), None);
-        assert_eq!(parse_auto_update_interval("garbage"), None);
-    }
-
-    #[test]
-    fn update_download_progress_bar_clears_line_on_finish() {
-        let (printer, buffer) = memory_printer();
-        let mut progress = UpdateDownloadProgress::new_for_terminal(printer, 13);
-
-        progress.add(13);
-        progress.finish();
-
-        let output = String::from_utf8(buffer.lock().unwrap().clone()).unwrap();
-        assert!(output.contains("100%"));
-        assert!(
-            output.ends_with("\r                                                            \r"),
-            "{output:?}"
-        );
-    }
-
-    #[test]
-    fn update_download_progress_spinner_clears_line_on_finish() {
-        let (printer, buffer) = memory_printer();
-        let mut progress = UpdateDownloadProgress::new_for_terminal(printer, -1);
-
-        progress.add(17);
-        progress.finish();
-
-        let output = String::from_utf8(buffer.lock().unwrap().clone()).unwrap();
-        assert!(output.contains("17B"));
-        assert!(output.ends_with("\r                                        \r"));
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn update_lock_nonblocking_returns_none_when_held() {
-        let dir = tempfile::tempdir().unwrap();
-        let _held = acquire_update_lock(dir.path(), true, true, None)
-            .unwrap()
-            .unwrap();
-
-        assert!(
-            acquire_update_lock(dir.path(), false, true, None)
-                .unwrap()
-                .is_none()
-        );
-    }
-
-    #[cfg(any(unix, windows))]
-    #[test]
-    fn update_lock_timeout_returns_when_held() {
-        let dir = tempfile::tempdir().unwrap();
-        let _held = acquire_update_lock(dir.path(), true, true, None)
-            .unwrap()
-            .unwrap();
-
-        let started_at = Instant::now();
-        let result = acquire_update_lock_with_timeout(
-            dir.path(),
-            true,
-            true,
-            None,
-            Duration::from_millis(10),
-        );
-        let elapsed = started_at.elapsed();
-
-        match result {
-            Err(err) => {
-                assert_eq!(
-                    err.to_string(),
-                    "timed out waiting for update lock after 10ms"
-                );
-            }
-            Ok(_) => panic!("expected held update lock to time out"),
-        }
-        assert!(elapsed < Duration::from_secs(1), "{elapsed:?}");
-    }
-
-    #[test]
-    fn update_lock_wait_timeout_uses_shorter_request_timeout() {
-        assert_eq!(
-            update_lock_wait_timeout(Some(Duration::from_millis(250))),
-            Duration::from_millis(250)
-        );
-        assert_eq!(
-            update_lock_wait_timeout(Some(Duration::from_secs(120))),
-            UPDATE_LOCK_WAIT_TIMEOUT
-        );
-        assert_eq!(update_lock_wait_timeout(None), UPDATE_LOCK_WAIT_TIMEOUT);
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn should_attempt_update_returns_false_when_lock_is_held() {
-        let dir = tempfile::tempdir().unwrap();
-        let _held = acquire_update_lock(dir.path(), true, true, None)
-            .unwrap()
-            .unwrap();
-
-        assert!(!should_attempt_update_in(dir.path(), Duration::ZERO).unwrap());
-    }
-
-    #[test]
-    fn should_attempt_update_uses_metadata_timestamp() {
-        let dir = tempfile::tempdir().unwrap();
-        update_last_attempt_time(dir.path(), SystemTime::now()).unwrap();
-
-        assert!(!should_attempt_update_in(dir.path(), Duration::from_secs(60)).unwrap());
-        assert!(should_attempt_update_in(dir.path(), Duration::ZERO).unwrap());
-    }
-
-    #[test]
-    fn update_temp_dir_is_removed_on_drop() {
-        let path = {
-            let dir = create_update_temp_dir().unwrap();
-            let path = dir.path().to_path_buf();
-            assert!(path.exists());
-            path
-        };
-
-        assert!(!path.exists());
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn update_temp_dir_is_private_on_unix() {
-        use std::os::unix::fs::PermissionsExt;
-
-        let dir = create_update_temp_dir().unwrap();
-
-        let mode = std::fs::metadata(dir.path()).unwrap().permissions().mode() & 0o777;
-        assert_eq!(mode, 0o700);
-    }
-
-    #[test]
-    fn test_unpack_zip_artifact_path_traversal() {
-        let tests = [
-            ("normal file", "fetch.exe", false),
-            ("path traversal with ..", "../escape.txt", true),
-            (
-                "deep path traversal",
-                "../../Windows/System32/malicious.dll",
-                true,
-            ),
-            ("absolute path", "C:/Windows/System32/malicious.dll", true),
-            (
-                "absolute unix path",
-                "/Windows/System32/malicious.dll",
-                true,
-            ),
-            ("backslash traversal", "..\\escape.txt", true),
-            (
-                "absolute backslash path",
-                "\\Windows\\System32\\bad.dll",
-                true,
-            ),
-            ("alternate data stream", "fetch.exe:ads", true),
-            ("reserved device name", "CON", true),
-            ("reserved device name with extension", "dir/NUL.txt", true),
-            (
-                "reserved device name with trailing dot",
-                "dir/COM1./fetch.exe",
-                true,
-            ),
-            (
-                "reserved device name with trailing space before extension",
-                "dir/NUL .txt",
-                true,
-            ),
-            ("all-dot windows component", "dir/.../fetch.exe", true),
-            ("trailing-dot component", "fetch.exe.", true),
-            ("trailing-space component", "fetch.exe ", true),
-        ];
-
-        for (name, filename, want_err) in tests {
-            let archive = create_zip(&[(filename, b"content".as_slice(), false)]);
-            let dir = tempfile::tempdir().unwrap();
-            let err = unpack_zip_artifact(dir.path(), &archive).err();
-            assert_eq!(err.is_some(), want_err, "{name}");
-            if !want_err {
-                assert!(dir.path().join(filename).exists(), "{name}");
-            }
-        }
-    }
-
-    #[test]
-    fn test_unpack_zip_artifact_explicit_directory_entry() {
-        let archive = create_zip(&[
-            ("bin/", b"".as_slice(), true),
-            ("bin/fetch.exe", b"content".as_slice(), false),
-        ]);
-        let dir = tempfile::tempdir().unwrap();
-
-        unpack_zip_artifact(dir.path(), &archive).unwrap();
-
-        assert!(dir.path().join("bin").join("fetch.exe").exists());
-    }
-
-    #[test]
-    fn test_unpack_zip_artifact_truncates_existing_regular_file() {
-        let archive = create_zip(&[("fetch.exe", b"short".as_slice(), false)]);
-        let dir = tempfile::tempdir().unwrap();
-        std::fs::write(dir.path().join("fetch.exe"), b"longer content").unwrap();
-
-        unpack_zip_artifact(dir.path(), &archive).unwrap();
-
-        assert_eq!(
-            std::fs::read_to_string(dir.path().join("fetch.exe")).unwrap(),
-            "short"
-        );
-    }
-
-    #[test]
-    fn test_unpack_zip_artifact_rejects_expanded_size_limit_and_removes_partial_binary() {
-        let readme = vec![b'a'; 10];
-        let binary = vec![b'b'; 10];
-        let archive = create_zip(&[
-            ("README.txt", readme.as_slice(), false),
-            ("fetch.exe", binary.as_slice(), false),
-        ]);
-        let dir = tempfile::tempdir().unwrap();
-
-        let err = unpack_zip_artifact_with_limits(
-            dir.path(),
-            &archive,
-            ArchiveExtractionLimits {
-                max_entries: 4,
-                max_uncompressed_bytes: 16,
-            },
-        )
-        .unwrap_err();
-
-        assert!(
-            err.to_string()
-                .contains("uncompressed data exceeded maximum allowed size"),
-            "{err}"
-        );
-        assert!(!dir.path().join("fetch.exe").exists());
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn test_unpack_artifact_path_traversal() {
-        let tests = [
-            ("normal file", "fetch", false),
-            ("path traversal with ..", "../escape.txt", true),
-            ("deep path traversal", "../../etc/passwd", true),
-            ("absolute path", "/etc/passwd", true),
-        ];
-
-        for (name, filename, want_err) in tests {
-            let archive = create_tar_gz(&[(filename, b"content".as_slice(), 0o644, false)]);
-            let dir = tempfile::tempdir().unwrap();
-            let err =
-                unpack_artifact(dir.path(), "fetch-v1.2.3-linux-amd64.tar.gz", &archive).err();
-            assert_eq!(err.is_some(), want_err, "{name}");
-            if !want_err {
-                assert!(dir.path().join(filename).exists(), "{name}");
-            }
-        }
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn test_unpack_artifact_explicit_directory_entry() {
-        let archive = create_tar_gz(&[
-            ("bin/", b"".as_slice(), 0o755, true),
-            ("bin/fetch", b"content".as_slice(), 0o755, false),
-        ]);
-        let dir = tempfile::tempdir().unwrap();
-
-        unpack_artifact(dir.path(), "fetch-v1.2.3-linux-amd64.tar.gz", &archive).unwrap();
-
-        assert!(dir.path().join("bin").join("fetch").exists());
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn test_unpack_artifact_truncates_existing_regular_file() {
-        let archive = create_tar_gz(&[
-            ("fetch", b"longer content".as_slice(), 0o755, false),
-            ("fetch", b"short".as_slice(), 0o755, false),
-        ]);
-        let dir = tempfile::tempdir().unwrap();
-
-        unpack_artifact(dir.path(), "fetch-v1.2.3-linux-amd64.tar.gz", &archive).unwrap();
-
-        assert_eq!(
-            std::fs::read_to_string(dir.path().join("fetch")).unwrap(),
-            "short"
-        );
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn test_unpack_tar_gz_artifact_masks_special_permission_bits() {
-        use std::os::unix::fs::PermissionsExt;
-
-        let archive = create_tar_gz(&[("fetch", b"content".as_slice(), 0o4755, false)]);
-        let dir = tempfile::tempdir().unwrap();
-
-        unpack_artifact(dir.path(), "fetch-v1.2.3-linux-amd64.tar.gz", &archive).unwrap();
-
-        let mode = std::fs::metadata(dir.path().join("fetch"))
-            .unwrap()
-            .permissions()
-            .mode()
-            & 0o7777;
-        assert_eq!(mode, 0o755);
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn test_unpack_tar_gz_artifact_rejects_expanded_size_limit_and_removes_partial_binary() {
-        let readme = vec![b'a'; 10];
-        let binary = vec![b'b'; 10];
-        let archive = create_tar_gz(&[
-            ("README.txt", readme.as_slice(), 0o644, false),
-            ("fetch", binary.as_slice(), 0o755, false),
-        ]);
-        let dir = tempfile::tempdir().unwrap();
-
-        let err = unpack_tar_gz_artifact_with_limits(
-            dir.path(),
-            &archive,
-            ArchiveExtractionLimits {
-                max_entries: 4,
-                max_uncompressed_bytes: 16,
-            },
-        )
-        .unwrap_err();
-
-        assert!(
-            err.to_string()
-                .contains("uncompressed data exceeded maximum allowed size"),
-            "{err}"
-        );
-        assert!(!dir.path().join("fetch").exists());
-    }
-
-    #[test]
-    fn test_unpack_artifact_dispatches_zip_by_suffix() {
-        let archive = create_zip(&[("fetch.exe", b"content".as_slice(), false)]);
-        let dir = tempfile::tempdir().unwrap();
-
-        unpack_artifact(dir.path(), "fetch-v1.2.3-windows-amd64.zip", &archive).unwrap();
-
-        assert_eq!(
-            std::fs::read_to_string(dir.path().join("fetch.exe")).unwrap(),
-            "content"
-        );
-    }
-
-    #[test]
-    fn test_unpack_artifact_rejects_unknown_suffix() {
-        let dir = tempfile::tempdir().unwrap();
-
-        let err =
-            unpack_artifact(dir.path(), "fetch-v1.2.3-plan9-amd64.bin", b"content").unwrap_err();
-
-        assert!(
-            err.to_string()
-                .contains("unsupported self-update archive format")
-        );
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn test_can_replace_file_read_only_file_writable_directory() {
-        use std::os::unix::fs::PermissionsExt;
-
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("fetch");
-        std::fs::write(&path, b"binary").unwrap();
-        std::fs::set_permissions(&path, std::fs::Permissions::from_mode(0o555)).unwrap();
-
-        assert!(can_replace_file(&path));
-    }
-
-    #[cfg(unix)]
-    fn create_tar_gz(entries: &[(&str, &[u8], u32, bool)]) -> Vec<u8> {
-        let mut data = Vec::new();
-        {
-            let encoder = GzEncoder::new(&mut data, Compression::default());
-            let mut builder = Builder::new(encoder);
-            for (name, content, mode, is_dir) in entries {
-                let mut header = Header::new_old();
-                let name_bytes = name.as_bytes();
-                header.as_old_mut().name[..name_bytes.len()].copy_from_slice(name_bytes);
-                header.set_mode(*mode);
-                header.set_size(if *is_dir { 0 } else { content.len() as u64 });
-                if *is_dir {
-                    header.set_entry_type(tar::EntryType::Directory);
-                }
-                header.set_cksum();
-                if *is_dir {
-                    builder.append(&header, &[][..]).unwrap();
-                } else {
-                    builder.append(&header, *content).unwrap();
-                }
-            }
-            builder.finish().unwrap();
-        }
-        data
-    }
-
-    fn create_zip(entries: &[(&str, &[u8], bool)]) -> Vec<u8> {
-        let cursor = std::io::Cursor::new(Vec::new());
-        let mut writer = zip::ZipWriter::new(cursor);
-        for (name, content, is_dir) in entries {
-            let options = zip::write::SimpleFileOptions::default()
-                .compression_method(zip::CompressionMethod::Deflated)
-                .unix_permissions(if *is_dir { 0o755 } else { 0o644 });
-            if *is_dir {
-                writer.add_directory(*name, options).unwrap();
-            } else {
-                writer.start_file(*name, options).unwrap();
-                writer.write_all(content).unwrap();
-            }
-        }
-        writer.finish().unwrap().into_inner()
-    }
-}
